fix some publishing
